+++ gazon-service/src/main/resources/db/changelog/v-1.5-selected-product/01-create-selectedProduct-table.sql
@@ -1,0 +1,7 @@
+
+CREATE TABLE selected_product
+(
+    id         BIGSERIAL PRIMARY KEY,
+    product_id BIGINT REFERENCES product (id) ON DELETE CASCADE NOT NULL,
+    count       INTEGER NOT NULL
+);