+++ gazon-service/src/test/java/com/gitlab/service/ProductServiceTest.java
@@ -12,8 +12,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.*;
@@ -60,17 +59,15 @@ class ProductServiceTest {
     @Test
     void should_update_product() {
         long id = 4L;
-        Product productToUpdateWith = new Product();
-        productToUpdateWith.setName("name1");
-        productToUpdateWith.setStockCount(1);
-        productToUpdateWith.setDescription("name");
-        productToUpdateWith.setIsAdult(true);
-        productToUpdateWith.setCode("name");
-        productToUpdateWith.setWeight(1L);
-        productToUpdateWith.setPrice(BigDecimal.ONE);
+        Product productToUpdateWith = generateProduct();
 
-        Product productBeforeUpdate = new Product(id, "n1", 1, null, "name", true, "name", 1L, BigDecimal.ONE);
-        Product productFromFuture = new Product(id, "name1", 1, null, "name", true, "name", 1L, BigDecimal.ONE);
+        Product productBeforeUpdate = new Product();
+        productBeforeUpdate.setId(id);
+        productBeforeUpdate.setName("old name");
+        productBeforeUpdate.setDescription("old");
+
+        Product productFromFuture = generateProduct();
+        productFromFuture.setId(id);
 
         when(productRepository.findById(id)).thenReturn(Optional.of(productBeforeUpdate));
         when(productRepository.save(productFromFuture)).thenReturn(productFromFuture);
@@ -83,14 +80,8 @@ class ProductServiceTest {
     @Test
     void should_not_update_product_when_entity_not_found() {
         long id = 4L;
-        Product productToUpdateWith = new Product();
-        productToUpdateWith.setName("name1");
-        productToUpdateWith.setStockCount(1);
-        productToUpdateWith.setDescription("name");
-        productToUpdateWith.setIsAdult(true);
-        productToUpdateWith.setCode("name");
-        productToUpdateWith.setWeight(1L);
-        productToUpdateWith.setPrice(BigDecimal.ONE);
+        Product productToUpdateWith = generateProduct();
+
 
         when(productRepository.findById(id)).thenReturn(Optional.empty());
 
@@ -120,26 +111,13 @@ class ProductServiceTest {
         verify(productRepository, never()).deleteById(anyLong());
     }
 
-    private List<Product> generateProducts() {
-        return List.of(
-                new Product(1L, "name1", 1, null, "name", true, "name", 1L, BigDecimal.ONE),
-                new Product(2L, "name2", 1, null, "name", true, "name", 1L, BigDecimal.ONE),
-                new Product(3L, "name3", 1, null, "name", true, "name", 1L, BigDecimal.ONE),
-                new Product(4L, "name4", 1, null, "name", true, "name", 1L, BigDecimal.ONE),
-                new Product(5L, "name5", 1, null, "name", true, "name", 1L, BigDecimal.ONE));
-    }
-
-    private Product generateProduct() {
-        return new Product(1L, "name1", 1, null, "name", true, "name", 1L, BigDecimal.ONE);
-    }
-
     @Test
     void should_not_updated_product_name_field_if_null() {
         long id = 1L;
-        Product productToUpdateWith = new Product(23L, "name1", 1, null, "name", true, "name", 1L, BigDecimal.ONE);
+        Product productToUpdateWith = generateProduct();
         productToUpdateWith.setName(null);
 
-        Product productBeforeUpdate = new Product(id, "n1", 1, null, "name", true, "name", 1L, BigDecimal.ONE);
+        Product productBeforeUpdate = generateProduct();
 
         when(productRepository.findById(id)).thenReturn(Optional.of(productBeforeUpdate));
         when(productRepository.save(productBeforeUpdate)).thenReturn(productBeforeUpdate);
@@ -147,16 +125,16 @@ class ProductServiceTest {
         Optional<Product> actualResult = productService.update(id, productToUpdateWith);
 
         verify(productRepository).save(productBeforeUpdate);
-        assertEquals(productBeforeUpdate, actualResult.orElse(null));
+        assertNotNull(actualResult.orElse(productBeforeUpdate).getName());
     }
 
     @Test
     void should_not_updated_product_stockCount_field_if_null() {
         long id = 1L;
-        Product productToUpdateWith = new Product(23L, "name1", 1, null, "name", true, "name", 1L, BigDecimal.ONE);
+        Product productToUpdateWith = generateProduct();
         productToUpdateWith.setStockCount(null);
 
-        Product productBeforeUpdate = new Product(id, "n1", 1, null, "name", true, "name", 1L, BigDecimal.ONE);
+        Product productBeforeUpdate = generateProduct();
 
         when(productRepository.findById(id)).thenReturn(Optional.of(productBeforeUpdate));
         when(productRepository.save(productBeforeUpdate)).thenReturn(productBeforeUpdate);
@@ -164,16 +142,16 @@ class ProductServiceTest {
         Optional<Product> actualResult = productService.update(id, productToUpdateWith);
 
         verify(productRepository).save(productBeforeUpdate);
-        assertEquals(productBeforeUpdate, actualResult.orElse(null));
+        assertNotNull(actualResult.orElse(productBeforeUpdate).getStockCount());
     }
 
     @Test
     void should_not_updated_product_description_field_if_null() {
         long id = 1L;
-        Product productToUpdateWith = new Product(23L, "name1", 1, null, "name", true, "name", 1L, BigDecimal.ONE);
+        Product productToUpdateWith = generateProduct();
         productToUpdateWith.setDescription(null);
 
-        Product productBeforeUpdate = new Product(id, "n1", 1, null, "name", true, "name", 1L, BigDecimal.ONE);
+        Product productBeforeUpdate = generateProduct();
 
         when(productRepository.findById(id)).thenReturn(Optional.of(productBeforeUpdate));
         when(productRepository.save(productBeforeUpdate)).thenReturn(productBeforeUpdate);
@@ -181,7 +159,35 @@ class ProductServiceTest {
         Optional<Product> actualResult = productService.update(id, productToUpdateWith);
 
         verify(productRepository).save(productBeforeUpdate);
-        assertEquals(productBeforeUpdate, actualResult.orElse(null));
+        assertNotNull(actualResult.orElse(productBeforeUpdate).getDescription());
+    }
+
+    private List<Product> generateProducts() {
+        return List.of(
+                generateProduct(1L),
+                generateProduct(2L),
+                generateProduct(3L),
+                generateProduct(4L),
+                generateProduct(5L));
+    }
+
+    private Product generateProduct(Long id) {
+        Product product = generateProduct();
+        product.setId(id);
+        return product;
+    }
+
+    private Product generateProduct() {
+        Product product = new Product();
+        product.setId(1L);
+        product.setName("name1");
+        product.setStockCount(2);
+        product.setDescription("name");
+        product.setIsAdult(true);
+        product.setCode("name");
+        product.setWeight(2L);
+        product.setPrice(BigDecimal.ONE);
+        return product;
     }
 
 }